+++ packages/producer/src/regression-harness-lambda-local.ts
@@ -46,29 +46,8 @@ import type {
   SerializableDistributedRenderConfig,
 } from "@hyperframes/aws-lambda";
 
-/** Inputs for {@link runLambdaLocalRender}. Same contract as `runDistributedSimulatedRender`. */
-export interface RunLambdaLocalInput {
-  projectDir: string;
-  tempRoot: string;
-  renderedOutputPath: string;
-  fps: 24 | 30 | 60;
-  /**
-   * Width/height from the fixture's renderConfig. Forwarded directly to
-   * the Lambda event so this mode catches drift if the handler ever
-   * starts honouring `Config.width/height` for canvas sizing rather
-   * than reading the composition's `data-width`/`data-height`. The
-   * `distributed-simulated` mode hardcodes 1920×1080 because it
-   * bypasses the event-serialization boundary; lambda-local goes
-   * through it, which is the whole point.
-   */
-  width: number;
-  height: number;
-  format: "mp4" | "mov" | "png-sequence";
-  codec?: "h264" | "h265";
-  chunkSize?: number;
-  maxParallelChunks?: number;
-  variables?: Record<string, unknown>;
-}
+export type { RunLambdaLocalInput } from "./regression-harness-lambda-local-types.js";
+import type { RunLambdaLocalInput } from "./regression-harness-lambda-local-types.js";
 
 const FAKE_BUCKET = "harness-lambda-local";
