+++ packages/producer/tsconfig.json
@@ -20,5 +20,11 @@
     }
   },
   "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist", "src/**/*.test.ts", "src/**/__test_utils__/**"]
+  "exclude": [
+    "node_modules",
+    "dist",
+    "src/**/*.test.ts",
+    "src/**/__test_utils__/**",
+    "src/regression-harness-lambda-local.ts"
+  ]
 }
